Re-route everything to avoid B.Cu cuts across diff pair
Switched the positions of the TX and RX LEDs to match which side of the
chip they're on. Significantly re-routed TX/RX. Added via stitching.

The ground planes are much more contiguous, except for U1 pad 5 which
connects to GND on the back with a via.

(Kieran's suggestion was to remove the lines on the back that cut
horizontally across the differential pair; I don't know if having
RXLED's sink next to the diff pair will be a problem.)

Re-version the card to "EC Card 2"